Moves GetBridgePorts for network to core/network so that we can use it in network config sources.
