+++ network/utils.go
@@ -5,7 +5,6 @@ package network
 
 import (
 	"net"
-	"path/filepath"
 )
 
 var netListen = net.Listen
@@ -21,33 +20,4 @@ func SupportsIPv6() bool {
 	}
 	ln.Close()
 	return true
-}
-
-// GetBridgePorts extracts and returns the names of all interfaces configured as
-// ports of the given bridgeName from the Linux kernel userspace SYSFS location
-// "<sysPath/<bridgeName>/brif/*". SysClassNetPath should be passed as sysPath.
-// Returns an empty result if the ports cannot be determined reliably for any
-// reason, or if there are no configured ports for the bridge.
-//
-// Example call: network.GetBridgePorts(network.SysClassNetPath, "br-eth1")
-func GetBridgePorts(sysPath, bridgeName string) []string {
-	portsGlobPath := filepath.Join(sysPath, bridgeName, "brif", "*")
-	// Glob ignores I/O errors and can only return ErrBadPattern, which we treat
-	// as no results, but for debugging we're still logging the error.
-	paths, err := filepath.Glob(portsGlobPath)
-	if err != nil {
-		logger.Debugf("ignoring error traversing path %q: %v", portsGlobPath, err)
-	}
-
-	if len(paths) == 0 {
-		return nil
-	}
-
-	// We need to convert full paths like /sys/class/net/br-eth0/brif/eth0 to
-	// just names.
-	names := make([]string, len(paths))
-	for i := range paths {
-		names[i] = filepath.Base(paths[i])
-	}
-	return names
 }
